+++ activesupport/lib/active_support.rb
@@ -36,6 +36,7 @@ require "active_support/core_ext/date_and_time/compatibility"
 module ActiveSupport
   extend ActiveSupport::Autoload
 
+  autoload :AttributeMethods
   autoload :Concern
   autoload :CodeGenerator
   autoload :ActionableError