+++ activemodel/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Port `ActiveModel::AttributeMethods` to `ActiveSupport::AttributeMethods`
+
+    *Sean Doyle*
+
 *   Port the `type_for_attribute` method to Active Model. Classes that include
     `ActiveModel::Attributes` will now provide this method. This method behaves
     the same for Active Model as it does for Active Record.